✨: Add delete command
