+++ libs/api/pdf-manager/service/feature/src/lib/commands/handlers/delete-pdf-manager.handler.ts
@@ -1,40 +1,18 @@
 import { HttpService } from '@nestjs/axios';
 import { CommandHandler, EventPublisher, ICommandHandler } from '@nestjs/cqrs';
-import { lastValueFrom, map, tap } from 'rxjs';
 import { PdfManagerServiceModel } from '../../models/pdf-manager-service-feature.model';
-import { GlobalKey } from '@conversation-catcher/api/pdf-manager/shared';
 //import {  } from "../events/delete-.event";
 import { DeletePdfCommand } from '../impl/delete-pdf-manager.command';
+import { MongoDBAccess } from '@conversation-catcher/api/pdf-manager/repository/data-access';
 
 @CommandHandler(DeletePdfCommand)
 export class DeletePdfHandler implements ICommandHandler<DeletePdfCommand> {
-  constructor(private publisher: EventPublisher,private repository: HttpService) {} //private repository: Repository,
+  constructor(
+    private publisher: EventPublisher,
+    private repository: MongoDBAccess
+  ) {}
 
   async execute({ id }: DeletePdfCommand) {
-    const url =
-      'https://data.mongodb-api.com/app/data-dtzbr/endpoint/data/v1/action/';
-    const action = 'deleteOne';
-    const data = JSON.stringify({
-      collection: 'PDF',
-      database: 'PDF',
-      dataSource: 'Cluster0',
-      filter: { id: id },
-    });
-
-    const config = {
-      method: 'post',
-      headers: {
-        'Content-Type': 'application/json',
-        'Access-Control-Request-Headers': '*',
-        'api-key': GlobalKey.key,
-      },
-    };
-    return await lastValueFrom(
-      this.repository.post(url + action, data, config).pipe(
-        tap((res) => console.log(res.status)),
-        map((res) => res.data)
-      )
-    );
-    return { id } as Partial<PdfManagerServiceModel>;
+    return this.repository.deletePDF(id);
   }
 }
